HW_6/requirements.txt
streamlit
transformers
python-dotenv
PyPDF2
requests

